fix(stdlib): fix issues when using LVGL TLSF memory pool to manage more than 4 GiB of memory

+++ tests/src/test_cases/libs/test_tjpgd.c
@@ -48,7 +48,7 @@ void test_tjpgd_1(void)
 
     TEST_ASSERT_EQUAL_SCREENSHOT("libs/jpg_1.png");
 
-    uint32_t mem_before = lv_test_get_free_mem();
+    size_t mem_before = lv_test_get_free_mem();
     for(uint32_t i = 0; i < 20; i++) {
         create_images();
